==> 1.1. determinar_género.ipynb <==
import pandas as pd
import os
import gender_guesser.detector as gender
import naiveBayesClassifier

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

# Leyendo base
df_contacts = pd.read_csv("undisclosed")

type(df_contacts.tel.iloc[0])

gd = gender.Detector()
df_contacts['genero_determinado_gender_guesser'] = df_contacts['nombre'].map(lambda x: gd.get_gender(x))

df_contacts

#from genderizer.genderizer import Genderizer
#df_contacts['genero_determinado_genderizer'] = df_contacts['genderizer'].apply(lambda x: Genderizer.detect(firstName=x, lang='es'))

df_contacts.to_excel('final_versions/df_contacts_extended_genderized.xlsx')
df_contacts.to_csv('final_versions/df_contacts_extended_genderized.csv')
