==> 1.2. arreglar_género.ipynb <==
import pandas as pd
import numpy as np
import os

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

# Leyendo base
df_contacts = pd.read_csv("final_versions/df_contacts_extended_genderized.csv")

df_contacts.head()

# Revisando forma de datos
df_contacts.genero_determinado_gender_guesser.value_counts()

# Revisando forma de datos
df_contacts.genero.value_counts()

# Modificando
df_contacts['genero'] = df_contacts['genero'].replace(['m', 'Masculino'], 'M')

df_contacts['genero_determinado_gender_guesser'] = df_contacts['genero_determinado_gender_guesser'].replace(['male', 'mostly_male'], 'M')
df_contacts['genero_determinado_gender_guesser'] = df_contacts['genero_determinado_gender_guesser'].replace(['female', 'mostly_female'], 'F')
df_contacts['genero_determinado_gender_guesser'] = df_contacts['genero_determinado_gender_guesser'].replace(['andy'], 'unknown')

# Replace empty cells in 'genero' with NaN
df_contacts['genero'] = df_contacts['genero'].replace('', np.nan)

# Fill NaN values in 'genero' with values from 'genero_determinado_gender_guesser' if not 'unknown'
df_contacts.loc[(df_contacts['genero'].isna()) & (df_contacts['genero_determinado_gender_guesser'] != 'unknown'), 'genero'] = df_contacts['genero_determinado_gender_guesser']

# Replace remaining NaN values in 'genero' with empty strings
df_contacts['genero'] = df_contacts['genero'].fillna('')

df_contacts.to_excel('final_versions/df_contacts_extended_genderized_arranged.xlsx')
df_contacts.to_csv('final_versions/df_contacts_extended_genderized_arranged.csv')

df_contacts.tel
