==> 1.3. phone_fixer.ipynb <==
import numpy as np
import pandas as pd
import os

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

# Leyendo base
df_contacts = pd.read_csv("final_versions/df_contacts.csv")

df_contacts.tel

type(df_contacts.tel[0])

# use str.endswith to identify rows ending with '.0' and not NaN
mask = (~pd.isna(df_contacts['tel'])) & (df_contacts['tel'].str.endswith('.0'))

# use str.slice to remove the final two characters from rows that meet the condition
df_contacts.loc[mask, 'tel'] = df_contacts.loc[mask, 'tel'].str.slice(stop=-2)

# add '+' sign to rows starting with '569' and not NaN
mask2 = (~pd.isna(df_contacts['tel'])) & (df_contacts['tel'].str.startswith('569'))
df_contacts.loc[mask2, 'tel'] = '+' + df_contacts.loc[mask2, 'tel']

df_contacts.tel

df_contacts.to_csv('final_versions/df_contacts_tel_fixed.csv')
df_contacts.to_excel('final_versions/df_contacts_tel_fixed.xlsx')
