==> 2.2. concatenando_dfs.ipynb <==
import pandas as pd
import os

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

df_contacts_1 = pd.read_csv("df_contacts_located_1.csv")
df_contacts_2 = pd.read_csv("df_contacts_located_2.csv")
df_contacts_3 = pd.read_csv("df_contacts_located_3.csv")
df_contacts_3 = pd.read_csv("df_contacts_located_4.csv")

df_contacts = pd.concat([df_contacts_1, df_contacts_2, df_contacts_3], ignore_index=True)

df_contacts.to_csv('final_versions/df_contacts.csv')
df_contacts.to_excel('final_versions/df_contacts.xlsx')
