==> 3.1. contando_palabras.ipynb <==
import pandas as pd
import os
from collections import Counter

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

# Leyendo base
df_contacts = pd.read_csv("final_versions/df_contacts_extended_genderized_arranged.csv")

df_contacts.columns

df_contacts.head()

def most_common_words(column, n):
    words = []
    for text in column:
        if isinstance(text, str):
            words.extend(text.split())
    word_counts = Counter(words)
    return word_counts.most_common(n)

top_words = most_common_words([x for x in df_contacts.profesion_procesada_corregida if x], 100)
print(top_words)
