==> 4.1. incorporando_contactos_existentes.ipynb <==
import pandas as pd
import os

# Estableciendo directorio
tar = os.chdir(r"undisclosed")

# Leyendo base
df_contacts_new = pd.read_csv("nueva_base/new_cep_database.csv")

df_contacts_new.head()

# Consiguiendo una lista con nombres originales de columnas
df_contacts_columns_original = list(df_contacts_new.columns.values)

df_contacts_columns_new = [
    'email',
    'nombre',
    'apellido',
    'fecha_nacimiento',
    'institucion',
    'profesion',
    'tel',
    'universidad',
    'carrera',
    'c1',
    'c2',
    'c3',
    'genero',
    'confirm_time'
]

# Renombrando columnas
dict_columns = dict(zip(df_contacts_columns_original, df_contacts_columns_new))
df_contacts_new = df_contacts_new.rename(columns=dict_columns)

df_contacts_new.head()

# Poniendo todos los emails en minúscula
df_contacts_new['email'] = df_contacts_new.email.str.lower()

len(df_contacts_new)

df_contacts_new = df_contacts_new.drop_duplicates(subset=['email'])
len(df_contacts_new)

# Leyendo base antigua
df_contacts = pd.read_csv("final_versions/df_contacts_genderized_arranged.csv")

# Filter df_contacts based on ids in df_contacts_new
df_contacts_in_old = df_contacts_new[df_contacts_new['email'].isin(df_contacts['email'])]
df_contacts_not_in_old = df_contacts_new[~df_contacts_new['email'].isin(df_contacts['email'])]

print(len(df_contacts_in_old))
print(len(df_contacts_not_in_old))

df_contacts_in_old = df_contacts_in_old.loc[:, ['email', 'c1']]
df_contacts_in_old = df_contacts_in_old.rename(columns = {'c1': 'c4'})
df_contacts_in_old.head()

print(len(df_contacts))

df_contacts_with_c4 = pd.merge(df_contacts, df_contacts_in_old, on='email', how = 'left')
print(len(df_contacts_with_c4))

df_contacts_with_c4.to_excel('final_versions/df_contacts_with_c4.xlsx')
df_contacts_with_c4.to_csv('final_versions/df_contacts_with_c4.csv')

df_contacts_not_in_old.to_csv('nueva_base/df_contacts_not_in_old.csv')
